added working get_bumpmap dimentions
